+++ socket2.c
@@ -28,7 +28,7 @@ int main() {
 
     char buffer[512];
     char sender_ip[16], receiver_ip[16];
-    uint8_t tunnel_id;
+    uint16_t tunnel_id;
     u_int8_t reached = 0;
 
     struct sockaddr_in addr;
